I have added a layer of dynamic functions to my static arrays for building the carton plot outline.
Positioning of lid and wall elements was corrected; now, the curve is back to being clockwise,
as it should have always been.

I have made the initial move redundant. It is up to each element path to begin with a move command.
In future, where a continuous curve is desired, I will reintroduce a move in the beginning,
and add a termination Z at the end, and dynamically strip all external component moves from the outline,
but only when this entire system is stable enough to be trusted with continuous curves in production.

We already have a sidewall style for the left and right walls of the carton, including glue tab.
I have now added a basewall style for the opposite end of the face to where the closing lid extends.
This helped me greatly in the above diagnostic maintenance, and it will help others too where
custom elements I don't know about are required to be added to the drawings my system produces.